tdf#100037 vml import: set fly-anchored AS_CHARs above fly's zOrder
Normally an AS_CHAR is the lowest in the heaven-layer zOrder,
but if it is inside a fly, it should be just above the fly's zOrder.

In order to get at the parent's properties,
I had to remove the stack abstraction
so I changed the stack to a vector.
That seemed a lot saner than popping and re-pushing.

make CppunitTest_sw_ooxmlexport18 \
    CPPUNIT_TEST_NAME=testTdf100037_inlineZOrder2
make CppunitTest_sw_ooxmlexport18 \
    CPPUNIT_TEST_NAME=testTdf100037_inlineZOrder3

Change-Id: Idc159e8203b3f304133a9b110c135e4d0f001dbc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168565
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins