Handle CPDB lib in configure/build system
In preparation of Biswadeep's upcoming GSoC changes (see [1])
to update the CPDB (Common Print Dialog Backends) support to the
latest version and use the CPDB C API, add a
corresponding `--enable-cpdb` autogen option
(disabled by default).

When CPDB is enabled, check for the cpdb-frontend
library, define `ENABLE_CPDB` and set the compiler
and linker flags for building the `CPDManager`
(in vcl/unx/generic/printer/cpdmgr.cxx).

Add checks for existing code using `CPDBManager`,
so it only gets used when when CPDB is actually enabled.

While the cpdb-frontend library is currently not
actually used in `CPDManager` yet (as the previous approach
tried to use some DBUs API directly), it will be in
Biswadeep's upcoming changes.

With this commit in place, an `--enable-cpdb` build
works with this additional local test change in place
to actually make use of the library:

    diff --git a/vcl/unx/generic/printer/cpdmgr.cxx b/vcl/unx/generic/printer/cpdmgr.cxx
    index 7dc17cede353..0c3b71d519db 100644
    --- a/vcl/unx/generic/printer/cpdmgr.cxx
    +++ b/vcl/unx/generic/printer/cpdmgr.cxx
    @@ -22,6 +22,8 @@
     #include <cstddef>
     #include <unistd.h>

    +#include <cpdb/cpdb-frontend.h>
    +
     #include <unx/cpdmgr.hxx>

     #include <osl/file.h>
    @@ -259,6 +261,7 @@ CPDManager* CPDManager::tryLoadCPD()
             }
         }
     #endif
    +    cpdbPrintFile(nullptr, nullptr);
         return pManager;
     }

[1] https://lists.freedesktop.org/archives/libreoffice/2024-May/091911.html

Change-Id: Ifc50d2cd6496346bea55c73cb703e3c2d3eb44b8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168656
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Reviewed-by: Biswadeep Purkayastha <bpdps95@hotmail.com>