tdf#156717 Fix "LibreOffice Community" in PDF metadata
These settings are available to set the brand when building LibreOffice:

1) Community flavor: yes / no
This is set via --disable-community-flavor, that defines
HAVE_FEATURE_COMMUNITY_FLAVOR, which is used in cui/source/dialogs/about.cxx.

2) Product name: A string (set via --with-product-name='name')
This defines PRODUCTNAME, which is used in many places in the code.

Not every LibreOffice build is a "LibreOffice Community" flavor. This
patch fixes this issue by using community brand only when variable
HAVE_FEATURE_COMMUNITY_FLAVOR is set.

Change-Id: Ib74450f380fed7c24222f88e389ddb5363bda3c1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168676
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>
Tested-by: Jenkins