tdf#161508: add another loop control hack
It beats me how to resolve the oscillation cleanly, when in a row
split mode, lcl_CalcMinRowHeight calculates a small value, but in
a non-split mode,  it returns a larger value  (which is expected),
and that resulted in recalc,  getting stuck forever in the nested
SwTabFrame::MakeAll.

So this puts an oscillation control here. The placement is mostly
heuristical.  It is hackish also in the sense that it only checks
the frame size and position,  but ignores the state;  so it might
turn out that it returns too early,  when a different combination
of flags was about to be attempted.

The unit test tests two things:
1. The main aspect of freeze;
2. The expected correct layout.

If (when) the hack turns out problematic,  and its fix happens to
break the second part of the test,  that is unfortunate,  but the
most important thing is to keep it from hanging.

Change-Id: If31d8527b4677b5211dcd3308578118c7066d68c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168677
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 0c49aa58cfbb81073e34b1d47861a5a1fdd44114)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168627
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168773
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>