tdf#161090 - Spec how many / which values are in remainder of-Pie chart
This commit moves bar-of-pie and pie-of-pie chart types into a separate
line in the chart type selection UI from 'plain' pie chart types. This
allows for separate UI controls, including a spin button added in this
commit for of-pie types to specify the number of entries to be included
in the composite wedge and broken out in the right-hand subchart. Per
discussion in this and related bugs, this may not be the long-term best
approach for specifying which entries go into the composite, and it
still only allows for the last entries in the composite. But it is a
step towards allowing greater control.

I've also changed the default number in the composite from 3 to 2, to
better match MSO.

I/O of the 'number in the composite wedge' parameter is not included
in this commit. Implementing that for ODF and OOXML perhaps should be
a separate bug or bugs.

Change-Id: If4afc1417ea94c15e86a9a4dfe967a6f8ecb7ca8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168690
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>