Related: tdf#161533 Test fixes for CJK systems
Previously, `CppunitTest_sw_txtexport` and `CppunitTest_sw_txtimport`
passed on most systems, but failed on fontconfig systems with CJK
fallback fonts available.

This change updates `CppunitTest_sw_txtexport` to remove the CJK
requirement.

This change also updates `CppunitTest_sw_txtimport` to temporarily
remove `SAL_NON_APPLICATION_FONT_USE=abort`, which was responsible for
the test failure.

This fixes the build on Ubuntu 24.04 for example.
I figured it might be useful to include this for long-term branches.

Change-Id: I0a91f5cccb367825adbcd0f93abbc3b8e9005698
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168752
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171881
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>