tdf#146769 menuBarManager: dispatch with KeyModifier Shift/Ctrl/Alt
If any of the modifier keys are held down while a menu item is selected,
dispatch the modifier keys along with the command.

There is not anything using this yet,
but (in an abandoned patch)
I wanted to let the "Quit" command know if KEY_SHIFT is pressed,
so I can imagine this ability could be very handy elsewhere.

The infrastructure is all there behind the scenes,
just this piece needed to be plugged in.

So instead of abandoning it completely,
let me put this piece in place so it is ready
for the next guy who could use it.

Change-Id: Iff45928872f8958bdb694771cae8dfb347c646ff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171843
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>