tdf#137335 calculate paragraph height in RTF/DOCX
TabStops and spaces are not included in height calculation of paragraphs in Word.
With a compatibility option this is now also set in RTF import.
But if spaces or/or tabstops are the only parts of a line their font size
is taken into account.
The fix 89e7341025b607491c90efdb74708e63d875c1e5%5E%21 from tdf#142404
has been removed because it broke formatting of follow lines.

Change-Id: Id28488438462114bd54c4f8bd15e2ada606c4192
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168788
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>