tdf#158658 sw: text formatting: make TabOverMargin even more crazy
... to better match Word's formatting.

The bugdoc has a special case where a right-aligned tab is positioned at
the right margin exactly, which causes the bFull condition in
SwTabPortion::PreFormat() to be true.

An obvious change to replace rInf.Width() - rInf.X() with
rInf.GetLineWidth() in the condition makes no difference because
m_pLastTab was already reset previously; instead, pass in the last tab
from Format(), which indicates that the right edge position of the
previous text portion was found via that tab.

Additionally, change the condition that checks for TabOverMargin to also
allow equal positions, i.e., exactly at the end of the paragraph.

Change-Id: I5d3b1c13eca0bffa640745d7c5f4113699f79cad
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168823
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>