annot: update object text when annotation text changes
When the annotation text changes we need to sync this with the SDR
object (if the object is showing the annotation text - like free
text annotation).

Also move annotation object searching to the annotation class, so
it can be reused in AnnotationWindow.

Also change uso of UNO XAnnotation in AnnotationWindow and use
concrete type (sdr::annotation::Anootation) instead.

Change-Id: I46db479ae8471c82ac58487a0125df15697245be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168977
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>