tdf#159835 sfx2: Move logic to populate sidebar menus to TabBar
Move the logic to populate the menus from
`SidebarController::PopulatePopupMenus` (called
by `SidebarController::ShowPopupMenu`) to
`TabBar::OnToolboxClicked`. The latter called
`SidebarController::ShowPopupMenu`.
(In a first step, take over the existing implementation
but leave further changes for following commits.)

`TabBar::OnToolboxClicked` already collects all the deck
data relevant for the menus and the menus are members of that
class, so it seems more straighforward to populate
the menus there right away.
The only information needed from `SidebarController`
is whether the sidebar is docked or not, so add a
new method `SidebarController::IsDocked` to retrieve that
information.

Rename `SidebarController::ShowPopupMenu` to
`SidebarController::ConnectMenuActivateHandlers` as
it only connects the signals now, and rename
`PopupMenuProvider` to `PopupMenuSignalConnectFunction`
accordingly.

This commit does some refactoring in preparation of
fixing tdf#159835, no change in behavior intended (yet).

Change-Id: I8236f2467239e6a2f485d468656e961478eeb09c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169005
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169075