tdf#145735 qt avmedia: Show audio placeholder for audio files
So far, `QtPlayer::createPlayerWindow` was unconditionally
creating a video widget.

However, audio-only media files can be used as well,
therefore handle that case, too and create a widget
holding an audio icon as a placeholder instead if
the media doesn't contain video.
(This is the same icon shown when using qt5 that
doesn't use QtMultimedia).

As described in

    commit a99575f04fa9c858bfcd996f037444135810d43f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Sat Jun 1 07:32:22 2024 +0200

        tdf#194504 qt avmedia: Don't wait for video frame if there's none

, `QMediaPlayer::hasVideo()` only returns a useful
result once loading media has finished. Therefore,
if the player is still in that state in
`QtPlayer::createPlayerWindow`, defer
creating the widget to when the media status
changes.

With this commit in place, opening an Impress presentation
that contains an audio file (like attachment 194504
from tdf#145735) and starting presentation mode now shows an
"audio icon" as placeholder as expected when using
the qt6 VCL plugin.

(This commit here makes that work for presentation
mode, while the above-mentioned commit already made
the icon show as expected for non-presentation mode.)

Change-Id: I1ff7e8b8659162a748abc3f97a8d2181375c0e7c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169009
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>