tdf#159835 sfx2 Connect SidebarController menu handlers earlier
Instead of connecting the SidebarController handlers
for handling activation of `TabBar` menu items
every time in `TabBar::OnToolboxClicked` (which is the
handler that gets called when clicking the menu button),
do so once in the `TabBar` ctor.

This also ensures that the handlers are connected
before the menu shows for the first time, which is
no longer true for the previous approach since

    commit f075fa01cb4f74185f13eb0a8d7f84cf1f47af49
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Aug 22 10:26:32 2023 +0200

        tdf#141101 tdf#101886 a11y: Restore previous focus on col/line popup close

Now, clicking on the sidebar menu and selecting the
"Undock" entry works to undock the first time, not only
when opening the menu for the second time.

Populating all menu entries before the menu gets
opened for the first time still needs to be implemented
separately.

Change-Id: I7e5d5e511c2ce2c7304e8f88676970d76e6319df
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169193
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
(cherry picked from commit 504e5501fce3f53d15b53f344c20e4404ff8f3db)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169140