tdf#161635 sw DoNotCaptureDrawObjsOnPage: fix handling of left padding
Open the bugdoc, the shape has an 5cm left padding for its text, but
only 4cm of that is visible in Writer.

Checking the shape, part of that is outside the page frame, so the first
1cm of the left padding is not visible, visually resulting in a 4cm left
padding in Writer, but not in Word.

Fix the problem by extending SwFlyFrame::MakePrtArea(), so in case the
shape is partially outside the page frame, then we make sure to increase
the left padding enough that the nominal (5cm) left padding is inside
the page frame.

With this, the text on the title page of the document is visually
centered also in Writer, even if not using an explicit paragraph
alignment.

Change-Id: I5dcbcf407ed7f12f0bc13820fa39621a76e23fe7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169186
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit 61692b82ae4ff62662509b5979a7aabc7d380678)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169134