tdf#161648 sc a11y: Set a11y name for border style button in sidebar
Set accessible name for the button to open the popup
to select a border style.

This is announced by Orca with the gtk3 VCL plugin
when the button gets focused as expected.

Setting accessible names for the items in the
popup still remains to be done, s.a. discussion
in pending Gerrit change [1] that's related.
(If that gets merged, setting a11y names should
probably be done in C++ code, not the corresponding
.ui file.)

[1] https://gerrit.libreoffice.org/c/core/+/169052

Change-Id: I200e160e64735e8701866428bf51e8b955dcc930
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169188
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>