tdf#159835 sfx2: Keep sidebar menu up to date
No longer update the menu for the menu button
in the sidebar in the handler that gets called
when the button gets clicked, but continually
keep the menu up to date instead.

This ensures that the menu is up-to-date when
it gets shown. Updating the menu in the button handler
is too late since

        commit f075fa01cb4f74185f13eb0a8d7f84cf1f47af49
        Author: Michael Weghorn <m.weghorn@posteo.de>
        Date:   Tue Aug 22 10:26:32 2023 +0200

            tdf#141101 tdf#101886 a11y: Restore previous focus on col/line popup close

, as the handler is now no more called before the menu
gets shown.

Do the update in method `TabBar::UpdateMenus` instead
and call that one whenever items change or the
sidebar gets docked/undocked, as menu entries
are shown for the items and for (un)docking.

This makes the menu show all entries again
when opened the first time after starting Writer,
which was no longer the case after the
above-mentioned commmit.

Change-Id: I9a7e341e88d2de5f34e8f85ebada3ff2ebf6b47d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169196
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>