uui,ucb: webdav-curl: display curl error message on connection failure
Show this in the dialog, which requires extending DAVException to store
2 strings for this case.

Store it in the Exception::Message member of
InteractiveNetworkConnectException, which appears to be unused.

Add another UI string ERRCODE_INET_CONNECT_MSG so that the text can be
shown optionally in the dialog.

Change-Id: Iaf139d77051e7480113eb4df82ef40eb498f329e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169278
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins
(cherry picked from commit b02eb998c1f05ccb2c1e82f4ba24779398bce72e)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169297