tdf#161658 Limit new selection overlay to Calc only
Previous patches [1] and [2] introduced a contrast white line to the selection overlay that was supposed to affect only Calc, but it ended up affecting other areas of LO, such as the star math editor as well as Writer.

This patch makes sure the changes only affect cell selection in Calc.

[1] dc243f0122ba656d2630e93bebfb84a2bfe4042a
[2] 3c0db898092c2cf6148c01f6c561acc199d484f5

Change-Id: Ie910120e4b71c55ad8c00a905e1204e291a711f9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169292
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>
Tested-by: Jenkins