tdf#161648 sc a11y: Set a11y names for items in border style control
Set accessible names for the items in the "Borders"
control in the Calc Sidebar, "Properties" deck, section
"Cell Appearance". Use the same text as is used for the tooltip.

With this in place, Orca now announces what each item
is about when using the gtk3 VCL plugin, instead of just saying
"Button".

Move the gla11y suppressions to the .false file, as they
have been dealt with now, see doc at [1].

[1] https://wiki.documentfoundation.org/Development/Accessibility

Change-Id: Id1856e4b116acad9bf6d14132bc6d9c91a01ae06
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169321
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169331