tdf#161646 sw page number wizard: fit into existing margins
Often page numbers are inserted when a document is finished,
and at that point it would be annoying to repaginate everything,
so offer this option of trying to fit the page number
into the existing margins.

I was tempted to enable it by default,
except that it would then mismatch any pages
that had header/footers created via other methods.

Problem:
-how to know which font is really needed: western, CTL, CJK?

This patch depends on NFC b7fce2b26ee4ba585544457adc742807a2129d2c

Change-Id: I285a99f8e2f12c87a8bdee20a9c86a1d79f80dd0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169827
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Justin Luth <jluth@mail.com>