makefile simplification: cleanup template packaging mess
the tmplate packaging code was full of copy'n'paste code with little
understanding on how the rules actually work/what the statements
actually do. (e.g. using separate calls to addprefix can be a matter
of taste, but adding the call to get the extra files each time is
pointless and just adds the same set of files multiple times).

Unifying the fixed list of files and standardizing the base path for the
relative filenames further allows to streamline the recipes

Change-Id: I814fdb6f6534a9d5046281c45b6da87110f2093e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169342
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>