tdf#144208 speedup doc with lots of redline(14)
SwXTextRange is heavily used. Reduce the allocation cost of
this class by de-pimpl'ing this class.

One strange thing I noticed is that the destructor of the Impl indicates
that deleting the bookmark via deleteMark() has to happen under the SolarMutex,
but there are two places in the existing code which violate this already.

Change-Id: I171a0d82d6d020d9a43808c424f308722770783f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169354
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>