Related: tdf#161652 sw, RTF paste: only keep used paragraph styles
When pasting from old enough Impress that doesn't have commit
afb4ea67463d9f0200dc6216cfd932aec0984c82 (tdf#161652 editeng, RTF copy:
only write used paragraph styles, 2024-06-20), it still happened that we
got many styles from an Impress slide's paragraph in Writer than just
the style of that paragraph itself.

The problem is that if we want to avoid problems with bad user input,
that has to be handled on the RTF paste / import side, not on the
producing side.

Fix the problem by filtering out unused paragraph styles also on the RTF
import (paste) side, in the IsNewDoc() case, which is the clipboard case
(not RTF file open).

With this, we attempt to filter out not needed paragraph styles both on
the import and export side.

Change-Id: Ic2c63e5f45245bb4296ec0d1a95558c459667e29
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169445
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins