LOK: Do a ".uno:DocumentStatus" "state changed" callback
... along with proper state changed callbacks for .uno:SlideMasterPage,
.uno:NormalMultiPaneGUI, .uno:NotesMode, and others.

The ".uno:DocumentStatus" payload is a notification for the callback
that it may need to collect the document status.

This allows to avoid the bond between the mentioned true UNO command
names, and the status collection: since we now send the status updates
for all of them, the document status collection would be performed for
each of them in the absence of a dedicated signal.

This re-implements commit bbd32b39eae24e99d135ed5c48586db8c9b0fb23,
without introducing a new callback enum value.

Change-Id: I942e17298196ff39ef8817a4cd59ad0fb8fa0da1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169468
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Mike Kaganski <mike.kaganski@collabora.com>