vcl: remove partial GIF image loading
This removes the code that handles the IO_PENDING and the graphic
"context" handling from GIF image format.

Partial loading of images complicates the image filter life cycle
a lot (many exceptions) and is not really needed so much today as
this was needed in the past. In most cases we load the whole image
in one pass anyway. Even loading from the network should be fast
enough to not cause issues for the user.

Most image filters don't even implement this like PNG and nobody
noticed for many years that this is not supported.

To handle IO_PENDING case it is probably better to load the
bitstream into memory and then load the whole image after that
in one pass. This can also be implemented all inside (Imp)Graphic
in a very straight forward way.

In addition some broken GIF files for testCVEs now fail to load
completely and those need to be moved from pass/ to fail/.

Change-Id: I5fb7004a4aff957da872bb3f5c66b61bf95f18d7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/165212
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins