tdf#161718 sw: fix background flys blocking footnotes
The problem is that the document has an unwanted page break on the
paragraph with the footnote.

The reason is that lcl_GetFootnoteLower() tries to evade flys, but
doesn't take into account that background flys (Wrap Through) should be
ignored.

(somehow regression from commit c303981cfd95ce1c3881366023d5495ae2edce97)

Change-Id: I02578f14644e232fac127142fe12801101f87f86
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169530
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>